# docs/04-技术架构/当前版本/V1-openapi.yaml
openapi: 3.0.3
info:
  title: DGOS V1 Logical API
  version: "1.0.0"
  description: |
    DGOS V1 逻辑契约的 HTTPS JSON + SSE 投影。桌面 IPC 必须保持相同的动作、字段、权限和事件语义。
    该文件同时登记已实现的本地 contract projection 与尚未具备真实桌面/Web E2E 证据的 draft operation。
security:
  - dgosSession: []
  - dgosApiKey: []
servers:
  - url: https://localhost
    description: Docker Compose Web projection (planned)
...
paths:
...
  /api/v1/identity/admin/bootstrap:
    post:
      tags: [identity]
      operationId: bootstrapAdmin
      security: []
      x-tag-feature: V1-FR-010
      x-tag-implementation-status: planned
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminBootstrapRequest' }
      responses:
        '201':
          description: First administrator created and session issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminSessionReceipt' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/identity/admin/login:
    post:
      tags: [identity]
      operationId: adminLogin
      security: []
      x-tag-feature: V1-FR-010
      x-tag-implementation-status: planned
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminLoginRequest' }
      responses:
        '200':
          description: Authenticated administrator session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminSessionReceipt' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/identity/admin/session:
    get:
      tags: [identity]
      operationId: getAdminSession
      x-tag-feature: V1-FR-010
      x-tag-implementation-status: planned
      security:
        - dgosSession: []
      responses:
        '200':
          description: Current administrator session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminSession' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [identity]
      operationId: renewAdminSession
      x-tag-feature: V1-FR-010
      x-tag-implementation-status: planned
      security:
        - dgosSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SessionRenewRequest' }
      responses:
        '200':
          description: Renewed administrator session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminSessionReceipt' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      operationId: logoutAdminSession
      x-tag-feature: V1-FR-010
      x-tag-implementation-status: planned
      security: [{ dgosSession: [] }]
      description: Authenticated current-session logout with CSRF, no target identifier and no fresh-session requirement. Revoke only caller session, clear its cookie and invalidate bound launch tickets; other devices remain unchanged. Invalid/expired caller may clear local cookie but cannot mutate another session.
      responses:
        '204': { description: Current session revoked and cookie cleared }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/identity/admin/sessions:
    get:
      operationId: listAdminSessions
      x-tag-feature: V1-FR-010
      x-tag-implementation-status: planned
      security: [{ dgosSession: [] }]
      description: Active administrator Session only; list only its principal devices/sessions. Reuses current administrator session authorization, no new API-key scope. No fresh authentication needed for own redacted list. No owner override, bearer sessionId, Cookie, credential, full IP or raw User-Agent. No-store and read audit; pagination order createdAt plus management ID is stable.
      parameters:
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { in: query, name: cursor, schema: { type: string, maxLength: 512 }, description: Opaque owner-bound pagination cursor, not authentication token }
      responses:
        '200':
          description: Own device sessions without authentication material
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminSessionList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/identity/admin/sessions/{sessionId}:
    delete:
      tags: [identity]
      operationId: revokeAdminSession
      description: Path sessionId is the NON-authenticating sessionManagementId from listAdminSessions, never the bearer credential currently used internally as sessionId. Requires active fresh administrator Session, CSRF and same-principal target lookup under transaction. Unknown/cross-owner returns 404 without enumeration. Revocation and audit/outbox atomic; same requestId returns same redacted receipt, already-revoked target is idempotent, other devices unchanged. Use DELETE /identity/admin/session for ordinary self logout.
      x-tag-feature: V1-FR-010
      x-tag-implementation-status: planned
      security:
        - dgosSession: []
      parameters:
        - { in: path, name: sessionId, required: true, schema: { type: string, pattern: '^sm_[A-Za-z0-9_-]{32,128}$' }, description: Non-authenticating management ID; legacy parameter name retained, bearer session IDs rejected }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequestEnvelope' }
      responses:
        '200':
          description: Revoked same-principal device session, no bearer material
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminSessionRevocationReceipt' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
...
  securitySchemes:
    appLaunchTicket:
      type: apiKey
      in: query
      name: launchTicket
      description: Opaque short-lived session/subject/app/package/instance-bound static ticket, redacted from logs; no bridge authority.
    dgosSession:
      type: apiKey
      in: cookie
      name: dgos_session
      description: Revocable DGOS administrator or subject session cookie.
    dgosApiKey:
      type: http
      scheme: bearer
      bearerFormat: DGOS-API-Key
      description: DGOS API Key; the secret is never returned after creation.
  responses:
...
    Unauthorized:
      description: Authentication missing, expired, or revoked
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Subject lacks the required capability or scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found within the caller's authorized scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Idempotency, version, reservation, or resource conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: Request failed schema or domain validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    RateLimited:
      description: Authentication or operation rate limit exceeded
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    ServiceUnavailable:
      description: Required service is unavailable; no protected side effect was committed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [errorKey, message, requestId]
      properties:
        errorKey: { type: string, description: Stable machine-readable error key }
        message: { type: string, description: Safe human-readable message without secrets }
        requestId: { type: string }
        retryable: { type: boolean }
        details: { type: object, additionalProperties: true, description: Redacted diagnostic details }
    AdminBootstrapRequest:
      allOf:
        - $ref: '#/components/schemas/RequestEnvelope'
        - type: object
          required: [displayName, credential]
          properties:
            displayName: { type: string, minLength: 1 }
            credential: { type: string, format: password, writeOnly: true }
            recoveryRef: { type: string, writeOnly: true, description: Secret boundary reference }
    AdminLoginRequest:
      allOf:
        - $ref: '#/components/schemas/RequestEnvelope'
        - type: object
          required: [principalHint, credential]
          properties:
            principalHint: { type: string, description: Opaque login hint; errors must remain non-enumerable }
            credential: { type: string, format: password, writeOnly: true }
    SessionRenewRequest:
      allOf:
        - $ref: '#/components/schemas/RequestEnvelope'
        - $ref: '#/components/schemas/VersionedRequest'
    AdminSessionReceipt:
      type: object
      required: [requestId, principalId, sessionId, state, expiresAt, sessionVersion]
      properties:
        requestId: { type: string }
        principalId: { type: string }
        sessionId: { type: string }
        state: { type: string, enum: [active, revoked, expired] }
        expiresAt: { type: string, format: date-time }
        sessionVersion: { type: string }
        authFreshUntil: { type: string, format: date-time }
    AdminSession:
      allOf:
        - $ref: '#/components/schemas/AdminSessionReceipt'
        - type: object
          properties:
            createdAt: { type: string, format: date-time }
            lastActiveAt: { type: string, format: date-time }
            deviceSummary: { type: object, additionalProperties: true }
    AdminSessionList:
      type: object
      additionalProperties: false
      required: [requestId, items]
      properties:
        requestId: { type: string }
        items: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/AdminSessionDevice' } }
        nextCursor: { type: string, description: Absent on last page }
    AdminSessionDevice:
      type: object
      additionalProperties: false
      required: [sessionManagementId, current, state, createdAt, expiresAt, sessionVersion, deviceSummary]
      properties:
        sessionManagementId: { type: string, pattern: '^sm_[A-Za-z0-9_-]{32,128}$', description: Stable opaque random management reference mapped server-side to same-owner session; cannot authenticate or derive credential }
        current: { type: boolean }
        state: { type: string, enum: [active, revoked, expired] }
        createdAt: { type: string, format: date-time }
        lastActiveAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        revokedAt: { type: string, format: date-time }
        sessionVersion: { type: string, pattern: '^[1-9][0-9]*$' }
        deviceSummary:
          type: object
          additionalProperties: false
          required: [host]
          properties:
            host: { type: string, enum: [web, macos, unknown] }
            label: { type: string, maxLength: 128, description: Sanitized coarse device/browser label, not raw User-Agent or asserted identity }
      description: Derive expired from server time even if stored state remains active. Device hints never confer authorization; missing historic hints use unknown, never fabricate device facts.
    AdminSessionRevocationReceipt:
      type: object
      additionalProperties: false
      required: [requestId, sessionManagementId, state, revokedAt, sessionVersion]
      properties:
        requestId: { type: string }
        sessionManagementId: { type: string, pattern: '^sm_[A-Za-z0-9_-]{32,128}$' }
        state: { type: string, enum: [revoked] }
        revokedAt: { type: string, format: date-time }
        sessionVersion: { type: string, pattern: '^[1-9][0-9]*$' }
...
    VersionedRequest:
      type: object
      required: [baseVersion]
      properties:
        baseVersion: { type: string }
...
    RequestEnvelope:
      type: object
      required: [requestId]
      properties:
        requestId: { type: string }
...